Access cache by reading and not writing
Write into the memory only to make sure the compiler "can't" optimise
reads by already knowing what's in the read position.

Tested and seems to work. However sometimes the position with less
access time is position 0. This makes sense since the remaining
variables of the program fit on the same page as array[0] and therefore
accessing the variables caches the lower part of the array.